ai-core/requirements.txt
# MillionX AI Core Requirements
# Hackathon-Optimized (Windows & Linux compatible)

# Core Data Processing
pandas>=2.0.0,<3.0.0
numpy>=1.24.0,<2.0.0

# Machine Learning
scikit-learn>=1.3.0,<2.0.0
statsforecast>=1.5.0  # For AutoARIMA, AutoETS
vaderSentiment>=3.3.2  # Sentiment analysis
networkx>=3.0  # Graph analytics

# Database
sqlalchemy>=2.0.23,<3.0.0
psycopg2-binary>=2.9.9  # PostgreSQL adapter
# Note: SQLite comes with Python

# Kafka Integration
kafka-python>=2.0.2,<3.0.0

# Environment & Configuration
python-dotenv>=1.0.0,<2.0.0

# Logging (built-in, but explicit for clarity)
# logging - standard library

# Agent Orchestration (Optional - for agent_workflow.py)
# langgraph==0.0.46
# langchain==0.1.7
# langchain-openai==0.0.5
# langchain-core==0.1.10
openai==1.6.1
pydantic==2.5.3